Revert "Check that tests are working with the right account by commenting teardown"
This reverts commit 9fabc88ca67ff09fe5e9fc48ff276cb5026a9867.

+++ test_client/tests/src/com/owncloud/android/lib/test_project/test/UploadFileTest.java
@@ -115,12 +115,12 @@ public class UploadFileTest extends ActivityInstrumentationTestCase2<TestActivit
 
 	@Override
 	protected void tearDown() throws Exception {
-		/*if (mUploadedFilePath != null) {
+		if (mUploadedFilePath != null) {
 			RemoteOperationResult removeResult = mActivity.removeFile(mUploadedFilePath);
 			if (!removeResult.isSuccess()) {
 				Utils.logAndThrow(LOG_TAG, removeResult);
 			}
-		}*/
+		}
 		super.tearDown();
 	}
